Update ScAdapterService to match NServiceBusWindowsService
