+++ src/Templates/ScAdapterService/WindowsService.cs
@@ -1,0 +1,15 @@
+using System.ServiceProcess;
+
+namespace ScAdapterService
+{
+    class WindowsService : ServiceBase
+    {
+        readonly Host host;
+
+        public WindowsService(Host host) => this.host = host;
+
+        protected override void OnStart(string[] args) => host.Start().GetAwaiter().GetResult();
+
+        protected override void OnStop() => host.Stop().GetAwaiter().GetResult();
+    }
+}
