+++ src/Tests/TemplateTests.ScAdapterService.approved.txt
@@ -1,7 +1,5 @@
-﻿ProgramService.cs => 
+﻿Host.cs => 
 using System;
-using System.Linq;
-using System.ServiceProcess;
 using System.Threading.Tasks;
 using NServiceBus;
 using NServiceBus.Logging;
@@ -9,90 +7,120 @@ using ServiceControl.TransportAdapter;
 
 namespace ScAdapterService
 {
-    class ProgramService : ServiceBase
+    class Host
     {
-        const string runAsServiceArg = "--run-as-service";
+        // TODO: optionally choose a custom logging library
+        // https://docs.particular.net/nservicebus/logging/#custom-logging
+        // LogManager.Use<TheLoggingFactory>();
+        static readonly ILog log = LogManager.GetLogger<Host>();
 
         ITransportAdapter adapter;
 
-        static ILog logger;
+        // TODO: give the adapter an appropriate name
+        public string AdapterName => "TransportAdapter.ScAdapterService";
 
-        static ProgramService()
-        {
-            //TODO: optionally choose a custom logging library
-            // https://docs.particular.net/nservicebus/logging/#custom-logging
-            // LogManager.Use<TheLoggingFactory>();
-            logger = LogManager.GetLogger<ProgramService>();
-        }
-
-        static void Main(string[] args)
-        {
-            using (var service = new ProgramService())
-            {
-                // pass argument at command line to run as a windows service
-                if (args.Contains(runAsServiceArg))
-                {
-                    Run(service);
-                    return;
-                }
-
-                Console.Title = "ScAdapterService";
-                Console.CancelKeyPress += (sender, e) => { service.OnStop(); };
-                service.OnStart(null);
-                Console.WriteLine("\r\nPress enter key to stop program\r\n");
-                Console.Read();
-                service.OnStop();
-            }
-        }
-
-        protected override void OnStart(string[] args)
-        {
-            AsyncOnStart().GetAwaiter().GetResult();
-        }
-
-        async Task AsyncOnStart()
+        public async Task Start()
         {
             try
             {
-                var adapterConfig = new TransportAdapterConfig<LearningTransport, LearningTransport>("TransportAdapter.WindowsService");
+                var adapterConfig = new TransportAdapterConfig<LearningTransport, LearningTransport>(AdapterName);
 
                 adapterConfig.CustomizeEndpointTransport(t =>
                 {
-                //TODO: Customize the endpoint-facing side of the adapter
-                //Use exactly same settings as in regular endpoints
-            });
+                    //TODO: Customize the endpoint-facing side of the adapter
+                    //Use exactly the same settings as in regular endpoints
+                });
 
                 adapterConfig.CustomizeServiceControlTransport(t =>
                 {
-                //TODO: Customize the ServiceControl-facing side of the adapter
-                //e.g. specify the same connection string as ServiceControl uses.
-            });
+                    //TODO: Customize the ServiceControl-facing side of the adapter
+                    //e.g. specify the same connection string as ServiceControl uses.
+                });
 
                 adapter = TransportAdapter.Create(adapterConfig);
 
-                await adapter.Start().ConfigureAwait(false);
+                await adapter.Start();
             }
             catch (Exception exception)
             {
-                Exit("Failed to start", exception);
+                FailFast("Failed to start", exception);
             }
         }
 
-        void Exit(string failedToStart, Exception exception)
+        public async Task Stop()
         {
-            logger.Fatal(failedToStart, exception);
-
-            //TODO: When using an external logging framework it is important to flush any pending entries prior to calling FailFast
-            // https://docs.particular.net/nservicebus/hosting/critical-errors#when-to-override-the-default-critical-error-action
-            Environment.FailFast(failedToStart, exception);
+            try
+            {
+                // TODO: perform any futher shutdown operations before or after stopping the adapter
+                await adapter?.Stop();
+            }
+            catch (Exception ex)
+            {
+                FailFast("Failed to stop correctly.", ex);
+            }
         }
 
-        protected override void OnStop()
+        void FailFast(string message, Exception exception)
         {
-            adapter?.Stop().GetAwaiter().GetResult();
+            try
+            {
+                log.Fatal(message, exception);
+
+                // TODO: when using an external logging framework it is important to flush any pending entries prior to calling FailFast
+                // https://docs.particular.net/nservicebus/hosting/critical-errors#when-to-override-the-default-critical-error-action
+            }
+            finally
+            {
+                Environment.FailFast(message, exception);
+            }
         }
     }
 }
+
+
+
+Program.cs => 
+using System;
+using System.Linq;
+using System.ServiceProcess;
+using System.Threading.Tasks;
+
+namespace ScAdapterService
+{
+    static class Program
+    {
+        // TODO: consider using C# 7.1 or later, which will allow
+        // removal of this method, and renaming of MainAsync to Main
+        public static void Main(string[] args) => MainAsync(args).GetAwaiter().GetResult();
+
+        public async static Task MainAsync(string[] args)
+        {
+            var host = new Host();
+
+            // pass this command line option to run as a windows service
+            if (args.Contains("--run-as-service"))
+            {
+                using (var windowsService = new WindowsService(host))
+                {
+                    ServiceBase.Run(windowsService);
+                    return;
+                }
+            }
+
+            Console.Title = host.AdapterName;
+
+            var tcs = new TaskCompletionSource<object>();
+            Console.CancelKeyPress += (sender, e) => { tcs.SetResult(null); };
+
+            await host.Start();
+            await Console.Out.WriteLineAsync("Press Ctrl+C to exit...");
+
+            await tcs.Task;
+            await host.Stop();
+        }
+    }
+}
+
 
 
 ScAdapterService.csproj => 
@@ -115,5 +143,24 @@ ScAdapterService.csproj =>
   </ItemGroup>
 
 </Project>
+
+
+WindowsService.cs => 
+using System.ServiceProcess;
+
+namespace ScAdapterService
+{
+    class WindowsService : ServiceBase
+    {
+        readonly Host host;
+
+        public WindowsService(Host host) => this.host = host;
+
+        protected override void OnStart(string[] args) => host.Start().GetAwaiter().GetResult();
+
+        protected override void OnStop() => host.Stop().GetAwaiter().GetResult();
+    }
+}
+
